Zadania 12-13 OPP.sda

+++ src/main/java/homework3_sda_OPP/Rectangle.java
@@ -11,6 +11,6 @@ public class Rectangle implements Figure {
 
     @Override
     public Double getArea() {
-        return length*width;
+        return (double) length*width;
     }
 }
